pts-core: Add RemoveDownloadFiles to user-config.xml for removing the originally downloaded files from a test once it has been installed
If RemoveDownloadFiles is set to TRUE within the user-config.xml file, once a test has been installed the originally downloaded files for that
test will be automatically removed. This conserves disk space but will cause these files to be re-downloaded the next time
the test needs to be re-installed. This will also not back up the downloaded files to the Phoronix Test Suite download
cache. Enabling this option is just recommended for users with very limited disk space.