install-sh: Add $DESTDIR support to make packaging easier
This significantly eases building packages of p-t-s.

If DESTDIR is unset, install-sh still installs as before

If DESTDIR is set, install-sh prefixes all install paths
with $DESTDIR.

Additionally, install-sh now tries to create the directory
$DESTDIR$INSTALL_PREFIX. This should not actually change
the install-sh behaviour except for avoiding the need to
create the directory before calling install-sh.