[pts-core] Added error message on failed installation in webui.
In webui, if a test fails to install, it will continue running the test because it does not run an initial check on the test run manager. I have fixed it so that run_benchmark ends if the check fails, and to send an error message to the web interface.