Create add_anchored_element function, add_element funcs [1/1]
[Problem]
The current strategy to create anchored element was to look for 'http_link' or 'xlink:href' attributes in the attributes
array. Creating a new anchored function is a cleaner way to approach this problem.

The add_element_* functions do not allow the user to specify a value or easily permit creating sub-elements.

[Solution]
An add_anchored_element function was written.

Additional add_element_* functions were created that allows the user to specify the element type, attributes, and a
value. There are also functions that support passing in an DOM element, which permits the creation of sub-elements. All
add_element_* functions return the DOM element.