Add polyline focus to line graph on hover [2/4]
[Problem]
Line graphs can be difficult to read. It was requested by a color blind phoronix reader to improve the readability of
these graphs.

[Solution]
Change the opacity of all other polylines to 0.0 on the hovering over a polyline. The opacities are reset when the mouse
leaves the polyline. The hover can be over the color box, key identifier, min/max/avg statistics, or the polyline.