Add thickness change to line graph on hover [1/1]
[Problem]
Line graphs can be difficult to read. It was requested by a color blind phoronix reader that highlighting of the lines
would help zir read the graphs.

[Solution]
Change the thickness and the opacity of the poly line on hover. The hover can be over the key identifier or the poly line.