Refactor graphic drawing code for cleanliness and flexiblity [3/4]
[Problem]
The gd code consisted of a single function with a giant switch statement that was really difficult to follow.
There was dead code in result_file_to_pdf.

[Solution]
The gd code to generate an image consisted of a single functiont that:
  1) Generated a gd object
  2) Converted the DOM elements into GD graphic objects
  3) Created the image file
Each of these steps was converted into a function.

Additionally, the gd drawing code for the different objects were converted into individual functions. Instead of using a
switch statement, an associative array between the DOM object type and the function that should be used to draw the
object is used.

The dead code in result_file_to_pdf was removed.