Add new command to download and check each test-profile
Checks each test profile to determine the status of the url download.
If successful, a check between the new download and the stored test-profile data
is performed to detect any differences in the filesize, MD5 and SHA256 checksums.

Current http status and any errors are output to a JSON file.
HTTP 302 redirections are also reported.

This test utilizes forking to prevent bottlenecks.

A file for a given url is downloaded once and reused for future tests thus preventing
multiple downloads of the same file.