fix removal of results not containing '1'
when refreshing, the 'Only show results matching....' field
gets evaluated to true because the check_request_for_var function
convert empty strings to true. The value is likely assumed to be a
string by the following code in process_request_to_attribute, and is
implicitely converted to '1'. Then, all results which don't match '1'
in their description, title, or ID are removed.