fix: is_proxy_setup() should return 'true' if the proxy is set
pts_network::is_proxy_setup() was introduced in e36e865 for the reason
"Don't use HTTPS on proxies when connecting to OpenBenchmarking.org",
however, in reality, HTTPS is only used if a proxy is configured.

is_proxy_setup() returns TRUE if no proxy is set, which seems the
opposite of the expected return value from the function name,
unlike other functions starting with is_*().