pts-core: Parsing result format with ":#_RESULT_#"
Some benchmark output reports are in the format:
xyz score:XYZ

Here there is no space between ':' and actual result.

To parse such format create a template something like this:
<OutputTemplate>xxx:#_RESULT_#</OutputTemplate>

This type of the format can't be parsed correctly.

The current implementation expects a "space" after column.
For example:
In test-profiles/system/selenium-1.0.18/results-definition.xml
<ResultsParser>
  <OutputTemplate>WEBXPRT SCORE RESULT: #_RESULT_#</OutputTemplate>

To be able to parse when there is no space between ":" and "#" add a
a space character in the template line.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>